+++ renderer/lib/locales.js
@@ -1,6 +1,7 @@
 //
-// https://www.electronjs.org/docs/api/locales
+// https://www.electronjs.org/docs/api/locales (dead link)
 //
+/*
 const locales =
 {
     "af": "Afrikaans",
@@ -136,6 +137,74 @@ const locales =
     "zh-CN": "Chinese (Simplified)",
     "zh-TW": "Chinese (Traditional)",
     "zu": "Zulu"
+};
+*/
+//
+// https://developer.chrome.com/docs/extensions/reference/i18n/
+// https://developer.chrome.com/docs/webstore/i18n/#choosing-locales-to-support
+// https://www.chromium.org/developers/design-documents/extensions/how-the-extension-system-works/i18n/
+//
+const locales =
+{
+    "am": "Amharic",
+    "ar": "Arabic",
+    "bg": "Bulgarian",
+    "bn": "Bengali",
+    "ca": "Catalan",
+    "cs": "Czech",
+    "da": "Danish",
+    "de": "German",
+    "el": "Greek",
+    "en": "English",
+    "en_GB": "English (Great Britain)",
+    "en_US": "English (USA)",
+    "es": "Spanish",
+    "es_419": "Spanish (Latin America and Caribbean)",
+    "et": "Estonian",
+    "fa": "Persian",    // ?
+    "fi": "Finnish",
+    "fil": "Filipino",
+    "fr": "French",
+    "gu": "Gujarati",
+    "he": "Hebrew",
+    "hi": "Hindi",
+    "hr": "Croatian",
+    "hu": "Hungarian",
+    "id": "Indonesian",
+    "it": "Italian",
+    "ja": "Japanese",
+    "kn": "Kannada",
+    "ko": "Korean",
+    "lt": "Lithuanian",
+    "lv": "Latvian",
+    "ml": "Malayalam",
+    "mr": "Marathi",
+    "ms": "Malay",  // ?
+    "nl": "Dutch",
+    "no": "Norwegian",  // ?
+    "pl": "Polish",
+    "pt_BR": "Portuguese (Brazil)",
+    "pt_PT": "Portuguese (Portugal)",
+    "ro": "Romanian",
+    "ru": "Russian",
+    "sk": "Slovak",
+    "sl": "Slovenian",
+    "sr": "Serbian",
+    "sv": "Swedish",
+    "sw": "Swahili",
+    "ta": "Tamil",
+    "te": "Telugu",
+    "th": "Thai",
+    "tr": "Turkish",
+    "uk": "Ukrainian",
+    "vi": "Vietnamese",
+    "zh_CN": "Chinese (China)",
+    "zh_TW": "Chinese (Taiwan)",
+    //
+    // "nb": "Norwegian (Bokmal)",
+    // "or": "Oriya",
+    // "pt": "Portuguese",
+    // "zh": "Chinese"
 };
 //
 module.exports = locales;