+++ renderer/units/JavaScript Runner/samples/Emoji/Display Emoji Test Pattern Names.js
@@ -1,3 +1,3 @@
 // Display Emoji Test Pattern Names
 const emojiTestPatterns = require ('./lib/unicode/emoji-test-patterns.js');
-return $.stringify (Object.keys (emojiTestPatterns), null, 4);
+$.write ($.stringify (Object.keys (emojiTestPatterns), null, 4));