+++ renderer/units/JavaScript Runner/samples/Unicode/Display Unicode Data Count.js
@@ -1,3 +1,3 @@
 // Display Unicode Data Count
 const unicodeData = require ('./lib/unicode/parsed-unicode-data.js');
-return `Unicode Data Count: ${Object.keys (unicodeData).length}`;
+$.write (`Unicode Data Count: ${Object.keys (unicodeData).length}`);