+++ renderer/units/JavaScript Runner/samples/Unicode/Display Unicode Extra Data.js
@@ -1,3 +1,3 @@
 // Display Unicode Extra Data
 const extraData = require ('./lib/unicode/parsed-extra-data.js');
-return $.stringify (extraData, null, 4);
+$.write ($.stringify (extraData, null, 4));